remove unnecessary `as usize`

+++ opentelemetry/src/sdk/trace/span_processor.rs
@@ -278,7 +278,7 @@ impl BatchSpanProcessor {
                     // Span batch interval time reached or a force flush has been invoked, export current spans.
                     BatchMessage::Flush(Some(ch)) => {
                         let mut results = Vec::with_capacity(
-                            spans.len() / config.max_export_batch_size + 1 as usize,
+                            spans.len() / config.max_export_batch_size + 1,
                         );
                         while !spans.is_empty() {
                             let batch = spans.split_off(
@@ -316,7 +316,7 @@ impl BatchSpanProcessor {
                     // Stream has terminated or processor is shutdown, return to finish execution.
                     BatchMessage::Shutdown(ch) => {
                         let mut results = Vec::with_capacity(
-                            spans.len() / config.max_export_batch_size + 1 as usize,
+                            spans.len() / config.max_export_batch_size + 1,
                         );
                         while !spans.is_empty() {
                             let batch = spans.split_off(
